Feature/dev db setting (#20)
* flyway 의존성 주입

* 자료형 int로 통합

* submodule changed

* test: 깨지는 테스트 수정

* test: 깨지는 테스트 수정

+++ doonut-core/src/main/java/com/doonutmate/doonut/image/entity/ImageEntity.java
@@ -37,7 +37,7 @@ public class ImageEntity extends BaseTimeEntity {
     int height;
 
     @Column
-    long width;
+    int width;
 
     @Column
     int capacity;
